requirements.txt
----------------------- -----------
albucore                0.0.24
annotated-types         0.7.0
anyio                   4.11.0
arrow                   1.4.0
asgiref                 3.9.1
asttokens               3.0.0
async-lru               2.0.5
attrs                   25.4.0
babel                   2.17.0
beautifulsoup4          4.14.2
bleach                  6.3.0
certifi                 2025.10.5
cffi                    2.0.0
charset-normalizer      3.4.4
colorama                0.4.6
comm                    0.2.3
contourpy               1.3.2
cycler                  0.12.1
debugpy                 1.8.17
decorator               5.2.1
defusedxml              0.7.1
Django                  5.2.6
executing               2.2.1
fastjsonschema          2.21.2
filelock                3.20.0
fonttools               4.58.0
fqdn                    1.5.1
fsspec                  2025.9.0
h11                     0.16.0
httpcore                1.0.9
idna                    3.11
ImageHash               4.3.2
imageio                 2.37.0
imutils                 0.5.4
ipython_pygments_lexers 1.1.1
jedi                    0.19.2
Jinja2                  3.1.6
joblib                  1.5.2
json5                   0.12.1
jsonpointer             3.0.0
jupyter_core            5.9.1
jupyterlab_pygments     0.3.0
jupyterlab_widgets      3.0.15
kiwisolver              1.4.8
lark                    1.3.1
lazy_loader             0.4
MarkupSafe              3.0.3
matplotlib              3.10.3
matplotlib-inline       0.2.1
mistune                 3.1.4
mpmath                  1.3.0
narwhals                2.10.0
nest-asyncio            1.6.0
networkx                3.5
numpy                   2.2.6
opencv-python           4.12.0.88
opencv-python-headless  4.12.0.88
packaging               25.0
pandas                  2.3.3
pandocfilters           1.5.1
parso                   0.8.5
pillow                  11.2.1
pip                     25.0.1
platformdirs            4.5.0
plotly                  6.3.1
prometheus_client       0.23.1
prompt_toolkit          3.0.52
psutil                  7.1.2
pure_eval               0.2.3
pycparser               2.23
pydantic                2.12.3
pydantic_core           2.41.4
pygame                  2.6.1
Pygments                2.19.2
pyparsing               3.2.3
python-dateutil         2.9.0.post0
python-json-logger      4.0.0
pytz                    2025.2
PyWavelets              1.9.0
pywinpty                3.0.2
PyYAML                  6.0.3
pyzmq                   27.1.0
referencing             0.37.0
requests                2.32.5
rfc3339-validator       0.1.4
rfc3986-validator       0.1.1
rfc3987-syntax          1.1.0
rpds-py                 0.28.0
scikit-image            0.25.2
scikit-learn            1.7.2
scipy                   1.16.2
seaborn                 0.13.2
Send2Trash              1.8.3
setuptools              80.9.0
simsimd                 6.5.3
six                     1.17.0
sniffio                 1.3.1
soupsieve               2.8
sqlparse                0.5.3
stack-data              0.6.3
stringzilla             4.2.3
sympy                   1.14.0
terminado               0.18.1
threadpoolctl           3.6.0
tifffile                2025.10.16
tinycss2                1.4.0
torch                   2.9.0
torchvision             0.24.0
tornado                 6.5.2
tqdm                    4.67.1
traitlets               5.14.3
typing_extensions       4.15.0
typing-inspection       0.4.2
tzdata                  2025.2
uri-template            1.3.0
urllib3                 2.5.0
wcwidth                 0.2.14
webcolors               24.11.1
webencodings            0.5.1
websocket-client        1.9.0
widgetsnbextension      4.0.14
